Renamed utils scripts and their imports
