+++ components/Charts/DrawChart.vue
@@ -29,7 +29,7 @@
 </template>
 
 <script>
-import { graphWinLoss } from '~/utils/graphWinLoss.js' 
+import { graphWinLoss } from '~/utils/chartWinLoss.js' 
 
 export default {
     name: "DrawChart",
