+++ components/Charts/ResByRating.vue
@@ -47,7 +47,7 @@
 </template>
 
 <script>
-import { graphResByOpp } from '~/utils/graphResByOpp.js';
+import { graphResByOpp } from '~/utils/chartResByOpp.js';
 
 export default {
   name: "ResByRating",
