+++ utils/chartElo.js
@@ -224,7 +224,6 @@ export function graphElo(timeClass="rapid")  {
 
   }
 
-  console.log(allData)
   return allData;
  }
